nok-controller/requirements.txt
kubernetes
python-consul